requirements.txt
appnope==0.1.0
astroid==2.2.5
backcall==0.1.0
confusable-homoglyphs==3.2.0
decorator==4.4.0
Django==2.2
django-redis==4.10.0
django-registration==3.0
factory-boy==2.11.1
Faker==1.0.4
gunicorn==19.9.0
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.17
jedi==0.13.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.1
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
wrapt==1.11.1
